# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/extras:
    get:
      operationId: listAdminExtras
      tags: [admin-extras]
      summary: List all extras (admin)
      description: Returns all extras including inactive ones.
      responses:
        "200":
          description: Array of all extras
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminExtra"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createAdminExtra
      tags: [admin-extras]
      summary: Create extra (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateExtraBody"
      responses:
        "201":
          description: Created extra
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminExtra"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/extras/{id}:
    get:
      operationId: getAdminExtra
      tags: [admin-extras]
      summary: Get single extra (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Extra record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminExtra"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Extra not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Admin Rates ──────────────────────────────────────────────────────────────

    patch:
      operationId: updateAdminExtra
      tags: [admin-extras]
      summary: Update extra (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateExtraBody"
      responses:
        "200":
          description: Updated extra
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminExtra"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Extra not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAdminExtra
      tags: [admin-extras]
      summary: Delete extra (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Extra not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/customers:
    get:
      operationId: listAdminCustomers
      tags: [admin-customers]
      summary: Paginated customer list (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated customer list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCustomerPaginatedResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createAdminCustomer
      tags: [admin-customers]
      summary: Create customer (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateCustomerBody"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCustomer"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/customers/{id}:
    get:
      operationId: getAdminCustomer
      tags: [admin-customers]
      summary: Get single customer (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCustomer"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Admin Bookings ───────────────────────────────────────────────────────────

    patch:
      operationId: updateAdminCustomer
      tags: [admin-customers]
      summary: Update customer (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateCustomerBody"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCustomer"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAdminCustomer
      tags: [admin-customers]
      summary: Delete customer (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/bookings:
    get:
      operationId: listAdminBookings
      tags: [admin-bookings]
      summary: Paginated booking list (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, CONFIRMED, DELIVERED, RETURNED, CANCELED, NO_SHOW]
        - name: paymentStatus
          in: query
          schema:
            type: string
            enum: [UNPAID, HALF, PAID, REFUNDED]
        - name: search
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated booking list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBookingPaginatedResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createAdminBooking
      tags: [admin-bookings]
      summary: Create booking (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateBookingBody"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBookingDetail"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/bookings/{id}:
    get:
      operationId: getAdminBooking
      tags: [admin-bookings]
      summary: Get single booking with detail (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking with nested entities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBookingDetail"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Admin Dashboard ───────────────────────────────────────────────────────────

    patch:
      operationId: updateAdminBooking
      tags: [admin-bookings]
      summary: Update booking (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateBookingBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBookingDetail"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAdminBooking
      tags: [admin-bookings]
      summary: Soft-delete booking (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/bookings/{id}/status:
    patch:
      operationId: updateAdminBookingStatus
      tags: [admin-bookings]
      summary: Update booking status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateBookingStatusBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBookingDetail"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Locations ──────────────────────────────────────────────────────────────
...
    AdminPaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
      required:
        - page
        - limit
        - total
...
    AdminExtra:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: string
        currency:
          type: string
        pricingType:
          type: string
          enum: [per_day, per_trip]
        maxDays:
          type: ["integer", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - price
        - currency
        - pricingType
        - isActive
        - createdAt
        - updatedAt

    # ─── Admin Rates ─────────────────────────────────────────────────────────────
...
    AdminCustomer:
      type: object
      properties:
        id:
          type: integer
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        fullName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - createdAt
        - updatedAt
    AdminCustomerPaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminCustomer"
        meta:
          $ref: "#/components/schemas/AdminPaginationMeta"
      required:
        - data
        - meta

    # ─── Admin Bookings ──────────────────────────────────────────────────────────
    AdminBookingRow:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [PENDING, CONFIRMED, DELIVERED, RETURNED, CANCELED, NO_SHOW]
        paymentStatus:
          type: string
          enum: [UNPAID, HALF, PAID, REFUNDED]
        contactFullName:
          type: string
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        pickupDatetime:
          type: string
          format: date-time
        dropoffDatetime:
          type: string
          format: date-time
        totalAmount:
          type: ["string", "null"]
        currency:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        broker:
          type: ["string", "null"]
        customer:
          type: object
          properties:
            id:
              type: integer
            fullName:
              type: ["string", "null"]
            email:
              type: ["string", "null"]
          required:
            - id
        vehicle:
          type: ["object", "null"]
          properties:
            id:
              type: integer
            licensePlate:
              type: ["string", "null"]
            modelName:
              type: ["string", "null"]
          required:
            - id
        pickupLocation:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
          required:
            - id
            - name
        dropoffLocation:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
          required:
            - id
            - name
        partner:
          type: ["object", "null"]
          properties:
            id:
              type: integer
            name:
              type: string
          required:
            - id
            - name
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - status
        - paymentStatus
        - contactFullName
        - pickupDatetime
        - dropoffDatetime
        - customer
        - pickupLocation
        - dropoffLocation
        - createdAt
    AdminBookingExtra:
      type: object
      properties:
        id:
          type: integer
        extraId:
          type: integer
        extraName:
          type: string
        quantity:
          type: ["integer", "null"]
        priceAtBooking:
          type: ["string", "null"]
      required:
        - id
        - extraId
        - extraName
    AdminBookingPayment:
      type: object
      properties:
        id:
          type: integer
        method:
          type: string
          enum: [CARD, CASH, BANK_TRANSFER, STRIPE, PAYPAL]
        status:
          type: string
          enum: [UNPAID, HALF, PAID, REFUNDED]
        amount:
          type: string
        currency:
          type: string
        transactionId:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - method
        - status
        - amount
        - currency
    AdminBookingDetail:
      allOf:
        - $ref: "#/components/schemas/AdminBookingRow"
        - type: object
          properties:
            userId:
              type: integer
            vehicleId:
              type: ["integer", "null"]
            vehicleGroupId:
              type: ["integer", "null"]
            vehicleModelId:
              type: ["integer", "null"]
            rateId:
              type: ["integer", "null"]
            rateTierId:
              type: ["integer", "null"]
            pricePerDay:
              type: ["string", "null"]
            baseRate:
              type: ["string", "null"]
            taxes:
              type: ["string", "null"]
            fees:
              type: ["string", "null"]
            discount:
              type: ["string", "null"]
            oneWayFee:
              type: ["string", "null"]
            deliveryFee:
              type: ["string", "null"]
            deposit:
              type: ["string", "null"]
            notes:
              type: ["string", "null"]
            documentType:
              type: ["string", "null"]
            documentNumber:
              type: ["string", "null"]
            pickupPhoto:
              type: ["string", "null"]
            returnPhoto:
              type: ["string", "null"]
            updatedAt:
              type: string
              format: date-time
            deletedAt:
              type: ["string", "null"]
            extras:
              type: array
              items:
                $ref: "#/components/schemas/AdminBookingExtra"
            payments:
              type: array
              items:
                $ref: "#/components/schemas/AdminBookingPayment"
          required:
            - userId
            - extras
            - payments
            - updatedAt
    AdminBookingPaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminBookingRow"
        meta:
          $ref: "#/components/schemas/AdminPaginationMeta"
      required:
        - data
        - meta

    # ─── Admin Dashboard ──────────────────────────────────────────────────────────
...
    DeleteResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ─── Admin Location Write Bodies ─────────────────────────────────────────────
...
    AdminCreateExtraBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: string
        currency:
          type: string
        pricingType:
          type: string
          enum: [per_day, per_trip]
        maxDays:
          type: integer
        isActive:
          type: boolean
      required:
        - name
        - price
        - currency
        - pricingType
    AdminUpdateExtraBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: string
        currency:
          type: string
        pricingType:
          type: string
          enum: [per_day, per_trip]
        maxDays:
          type: integer
        isActive:
          type: boolean

    # ─── Admin Rates Write Bodies ────────────────────────────────────────────────
...
    AdminCreateCustomerBody:
      type: object
      properties:
        email:
          type: string
        phone:
          type: string
        fullName:
          type: string
    AdminUpdateCustomerBody:
      type: object
      properties:
        email:
          type: string
        phone:
          type: string
        fullName:
          type: string

    # ─── Admin Bookings Write Bodies ─────────────────────────────────────────────
    AdminCreateBookingBody:
      type: object
      properties:
        customerId:
          type: integer
          description: Existing customer ID. If omitted, customerData creates a new customer.
        customerData:
          type: object
          properties:
            fullName:
              type: string
            email:
              type: string
            phone:
              type: string
        contactFullName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        vehicleId:
          type: integer
        vehicleModelId:
          type: integer
        vehicleGroupId:
          type: integer
        pickupLocationId:
          type: integer
        dropoffLocationId:
          type: integer
        pickupDatetime:
          type: string
          format: date-time
        dropoffDatetime:
          type: string
          format: date-time
        status:
          type: string
          enum: [PENDING, CONFIRMED, DELIVERED, RETURNED, CANCELED, NO_SHOW]
        paymentStatus:
          type: string
          enum: [UNPAID, HALF, PAID, REFUNDED]
        rateId:
          type: integer
        rateTierId:
          type: integer
        pricePerDay:
          type: string
        totalAmount:
          type: string
        currency:
          type: string
        source:
          type: string
        broker:
          type: string
        notes:
          type: string
        documentType:
          type: string
        documentNumber:
          type: string
        deposit:
          type: string
      required:
        - contactFullName
        - pickupLocationId
        - dropoffLocationId
        - pickupDatetime
        - dropoffDatetime
    AdminUpdateBookingBody:
      type: object
      properties:
        contactFullName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        vehicleId:
          type: integer
        vehicleModelId:
          type: integer
        vehicleGroupId:
          type: integer
        pickupLocationId:
          type: integer
        dropoffLocationId:
          type: integer
        pickupDatetime:
          type: string
          format: date-time
        dropoffDatetime:
          type: string
          format: date-time
        status:
          type: string
          enum: [PENDING, CONFIRMED, DELIVERED, RETURNED, CANCELED, NO_SHOW]
        paymentStatus:
          type: string
          enum: [UNPAID, HALF, PAID, REFUNDED]
        rateId:
          type: integer
        rateTierId:
          type: integer
        pricePerDay:
          type: string
        totalAmount:
          type: string
        currency:
          type: string
        source:
          type: string
        broker:
          type: string
        notes:
          type: string
        documentType:
          type: string
        documentNumber:
          type: string
        deposit:
          type: string
    AdminUpdateBookingStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [PENDING, CONFIRMED, DELIVERED, RETURNED, CANCELED, NO_SHOW]
        paymentStatus:
          type: string
          enum: [UNPAID, HALF, PAID, REFUNDED]
      required:
        - status

    # ─── Admin Team ───────────────────────────────────────────────────────────────
...